+++ common/lib/xmodule/xmodule/modulestore/mongo/__init__.py
@@ -1,0 +1,5 @@
+from xmodule.modulestore.mongo.base import MongoModuleStore, MongoKeyValueStore, MongoUsage
+
+# Backwards compatibility for prod systems that refererence
+# xmodule.modulestore.mongo.DraftMongoModuleStore
+from xmodule.modulestore.mongo.draft import DraftModuleStore as DraftMongoModuleStore
